+++ examples/drafts-for-rfc.py
@@ -1,0 +1,53 @@
+# Copyright (C) 2020 University of Glasgow
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions
+# are met:
+#
+# 1. Redistributions of source code must retain the above copyright notice,
+#    this list of conditions and the following disclaimer.
+#
+# 2. Redistributions in binary form must reproduce the above copyright
+#    notice, this list of conditions and the following disclaimer in the
+#    documentation and/or other materials provided with the distribution.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+
+import os
+import sys
+
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+from datetime                 import datetime, timedelta
+from pathlib                  import Path
+from ietfdata.rfcindex        import *
+from ietfdata.datatracker     import *
+from ietfdata.datatracker_ext import *
+
+# =============================================================================
+
+dt = DataTrackerExt(cache_dir=Path("cache"))
+ri = RFCIndex()
+
+for rfcnum in ["RFC5801", "RFC5802", "RFC8256"]:
+    rfc = ri.rfc(rfcnum)
+    if rfc is not None:
+        for d in dt.drafts_for_rfc(rfc):
+            print(F"{d.rfc.doc_id} {d.draft.name}-{d.rev} {d.date} ", end="")
+            if d.submission is not None:
+                print(d.submission.resource_uri)
+            else:
+                print("(none)")
+    print("")
+
+# =============================================================================